✨ feat: 댓글 자신의 것이 아닐 시 삭제 불가 기능 추가

+++ src/main/java/com/example/harumeonglog/domain/comment/service/CommentCommandServiceImpl.java
@@ -3,6 +3,7 @@ package com.example.harumeonglog.domain.comment.service;
 import com.example.harumeonglog.domain.comment.dto.request.CommentRequest;
 import com.example.harumeonglog.domain.comment.entity.Comment;
 import com.example.harumeonglog.domain.comment.repository.CommentRepository;
+import com.example.harumeonglog.domain.member.entity.Member;
 import com.example.harumeonglog.global.error.code.CommentErrorCode;
 import com.example.harumeonglog.global.error.exception.CommentException;
 import lombok.RequiredArgsConstructor;
@@ -32,8 +33,13 @@ public class CommentCommandServiceImpl implements CommentCommandService {
     }
 
     @Override
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, Member member) {
         Comment comment = commentRepository.findById(commentId).orElseThrow(() -> new CommentException(CommentErrorCode.NOT_FOUND));
+
+        if (comment.getMember().equals(member)) {
+            throw new CommentException(CommentErrorCode.NOT_OWN);
+        }
+
         comment.delete();
     }
